Moves masp events data to core
