+++ crates/core/src/masp.rs
@@ -14,6 +14,7 @@ use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 
 use crate::address::{Address, DecodeError, HASH_HEX_LEN, MASP};
+use crate::hash::Hash;
 use crate::impl_display_and_from_str_via_format;
 use crate::storage::Epoch;
 use crate::string_encoding::{
@@ -530,5 +531,34 @@ impl FromStr for MaspValue {
             .or_else(|_err| {
                 ExtendedViewingKey::from_str(s).map(Self::FullViewingKey)
             })
+    }
+}
+
+/// Reference to a masp transaction inside a [`BatchedTx`]
+#[derive(Clone, Serialize, Deserialize)]
+pub struct MaspTxRef {
+    // FIXME: actually, are we using the commitment? Probably if I give the
+    // masp section ref I don't need the commitment anymore right?
+    /// The inner tx commitment's hash
+    pub cmt: Hash,
+    /// The hash of the masp [`Section::MaspTx`]
+    pub masp_section_ref: Hash,
+}
+
+/// The masp transactions' references of a given batch
+#[derive(Default, Clone, Serialize, Deserialize)]
+pub struct BatchMaspTxRefs(pub Vec<MaspTxRef>);
+
+impl Display for BatchMaspTxRefs {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", serde_json::to_string(self).unwrap())
+    }
+}
+
+impl FromStr for BatchMaspTxRefs {
+    type Err = serde_json::Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        serde_json::from_str(s)
     }
 }
