+++ crates/node/src/shell/finalize_block.rs
@@ -4,7 +4,6 @@ use data_encoding::HEXUPPER;
 use masp_primitives::merkle_tree::CommitmentTree;
 use masp_primitives::sapling::Node;
 use namada::core::storage::{BlockResults, Epoch, Header};
-use namada::events::extend::ValidMaspTxs;
 use namada::events::Event;
 use namada::gas::event::GasUsed;
 use namada::governance::pgf::inflation as pgf_inflation;
@@ -17,6 +16,7 @@ use namada::ledger::gas::GasMetering;
 use namada::ledger::ibc;
 use namada::ledger::pos::namada_proof_of_stake;
 use namada::ledger::protocol::DispatchError;
+use namada::masp::BatchMaspTxRefs;
 use namada::proof_of_stake;
 use namada::proof_of_stake::storage::{
     find_validator_by_raw_hash, write_last_block_proposer_address,
@@ -824,7 +824,7 @@ impl<'finalize> TempTxLogs {
     fn check_inner_results(
         &mut self,
         tx_result: &namada::tx::data::TxResult<protocol::Error>,
-        masp_tx_refs: ValidMaspTxs,
+        masp_tx_refs: BatchMaspTxRefs,
         tx_header: &namada::tx::Header,
         tx_index: usize,
         height: BlockHeight,
